fix: avoid rounding problems with microseconds
