+++ tests/unit/test_dtypes.py
@@ -621,51 +621,3 @@ def test_date_sub():
     do = pd.Series([pd.DateOffset(days=i) for i in range(4)])
     expect = dates.astype("object") - do
     assert np.array_equal(dates - do, expect)
-
-
-@pytest.mark.parametrize(
-    "value, expected", [("1", datetime.time(1)), ("1:2", datetime.time(1, 2))],
-)
-def test_short_time_parsing(value, expected):
-    assert _cls("time")([value])[0] == expected
-
-
-@pytest.mark.parametrize(
-    "value, error",
-    [
-        ("thursday", "Bad time string: 'thursday'"),
-        ("1:2:3thursday", "Bad time string: '1:2:3thursday'"),
-        ("1:2:3:4", "Bad time string: '1:2:3:4'"),
-        ("1:2:3.f", "Bad time string: '1:2:3.f'"),
-        ("1:d:3", "Bad time string: '1:d:3'"),
-        ("1:2.3", "Bad time string: '1:2.3'"),
-        ("", "Bad time string: ''"),
-        ("1:2:99", "second must be in 0[.][.]59"),
-        ("1:99", "minute must be in 0[.][.]59"),
-        ("99", "hour must be in 0[.][.]23"),
-    ],
-)
-def test_bad_time_parsing(value, error):
-    with pytest.raises(ValueError, match=error):
-        _cls("time")([value])
-
-
-@pytest.mark.parametrize(
-    "value, error",
-    [
-        ("thursday", "Bad date string: 'thursday'"),
-        ("1-2-thursday", "Bad date string: '1-2-thursday'"),
-        ("1-2-3-4", "Bad date string: '1-2-3-4'"),
-        ("1-2-3.f", "Bad date string: '1-2-3.f'"),
-        ("1-d-3", "Bad date string: '1-d-3'"),
-        ("1-3", "Bad date string: '1-3'"),
-        ("1", "Bad date string: '1'"),
-        ("", "Bad date string: ''"),
-        ("2021-2-99", "day is out of range for month"),
-        ("2021-99-1", "month must be in 1[.][.]12"),
-        ("10000-1-1", "year 10000 is out of range"),
-    ],
-)
-def test_bad_date_parsing(value, error):
-    with pytest.raises(ValueError, match=error):
-        _cls("date")([value])
